Remove Converting transform and step.
